Simplify GitLab CI example
The documented GitLab configuration is over-complicated, as it uses three containers (`docker:latest` and `gradle:3.4`, `docker:dind`).
The `docker:latest` container is unnecessary, and mounting the socket can be replaced by setting the `DOCKER_HOST` environment variable.
It took me a while to figure this out, so I would like to share my solution with the community.
The project that uses this new configuration is private. But I could set up an example on gitlab.com to show that it works.